Creating, Planning and Running a User Group v1
This is the first version of the guide 'Creating, Planning and Running a User Group' and guides readers through the thought process of what to consider when creating a user group. It includes a calendar of things to consider in the first months of starting a user group. It has space to journal and log ideas and progress along your journey.